pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "realtime-speech-coach"
version = "0.1.0"
description = "Realtime speech coach with transcription and AI suggestions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.30.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.1",
    "webrtcvad-wheels>=2.0.14",
    "faster-whisper>=1.0.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]